refactor(ext/node): consolidate fs modules (batch 2) (#32505)
## Summary

Continues the consolidation started in #32502 by inlining 5 more small
`_fs` modules into `fs.ts`:

- **`_fs_fsync.ts`** → `fsync`, `fsyncSync` inlined into `fs.ts`
- **`_fs_fdatasync.ts`** → `fdatasync`, `fdatasyncSync` inlined into
`fs.ts`
- **`_fs_link.ts`** → `link`, `linkSync` inlined into `fs.ts`
- **`_fs_unlink.ts`** → `unlink`, `unlinkSync` inlined into `fs.ts`
- **`_fs_fchmod.ts`** → `fchmod`, `fchmodSync` inlined into `fs.ts`

Additional changes:
- **`internal/fs/promises.ts`**: inlined `link`/`unlink` promise
versions
- **`internal/fs/handle.ts`**: imports `fchmod`/`fdatasync`/`fsync` from
`node:fs` and creates promisified versions locally
- **`cp/cp.ts`** and **`cp/cp_sync.ts`**: use
`Deno.remove`/`Deno.removeSync` directly instead of importing from the
now-deleted `_fs_unlink.ts`
- **Tests**: consolidated from 4 separate test files into `fs_test.ts`

This removes 5 small modules from the module graph, continuing to reduce
startup overhead.
