+++ ext/node/lib.rs
@@ -382,17 +382,13 @@ deno_core::extension!(deno_node,
     "_fs/cp/cp_sync.ts",
     "_fs/_fs_dir.ts",
     "_fs/_fs_exists.ts",
-    "_fs/_fs_fchmod.ts",
     "_fs/_fs_fchown.ts",
-    "_fs/_fs_fdatasync.ts",
     "_fs/_fs_fstat.ts",
-    "_fs/_fs_fsync.ts",
     "_fs/_fs_ftruncate.ts",
     "_fs/_fs_futimes.ts",
     "_fs/_fs_glob.ts",
     "_fs/_fs_lchmod.ts",
     "_fs/_fs_lchown.ts",
-    "_fs/_fs_link.ts",
     "_fs/_fs_lstat.ts",
     "_fs/_fs_lutimes.ts",
     "_fs/_fs_mkdir.ts",
@@ -412,7 +408,6 @@ deno_core::extension!(deno_node,
     "_fs/_fs_statfs.ts",
     "_fs/_fs_symlink.ts",
     "_fs/_fs_truncate.ts",
-    "_fs/_fs_unlink.ts",
     "_fs/_fs_utimes.ts",
     "_fs/_fs_watch.ts",
     "_fs/_fs_write.ts",
