+++ ext/node/polyfills/_fs/cp/cp.ts
@@ -7,7 +7,6 @@ import { mkdirPromise } from "ext:deno_node/_fs/_fs_mkdir.ts";
 import { opendirPromise } from "ext:deno_node/_fs/_fs_opendir.ts";
 import { readlinkPromise } from "ext:deno_node/_fs/_fs_readlink.ts";
 import { symlinkPromise } from "ext:deno_node/_fs/_fs_symlink.ts";
-import { unlinkPromise } from "ext:deno_node/_fs/_fs_unlink.ts";
 import { utimesPromise } from "ext:deno_node/_fs/_fs_utimes.ts";
 import { EEXIST, EINVAL, EISDIR, ENOTDIR } from "node:constants";
 import {
@@ -284,7 +283,7 @@ async function mayCopyFile(
   opts: CopyOptions,
 ) {
   if (opts.force) {
-    await unlinkPromise(dest);
+    await Deno.remove(dest);
     return _copyFile(srcStat, src, dest, opts);
   } else if (opts.errorOnExist) {
     throw new ERR_FS_CP_EEXIST({
@@ -450,6 +449,6 @@ async function onLink(
 }
 
 async function copyLink(resolvedSrc: string, dest: string): Promise<void> {
-  await unlinkPromise(dest);
+  await Deno.remove(dest);
   return symlinkPromise(resolvedSrc, dest);
 }
