+++ ext/node/polyfills/fs.ts
@@ -16,16 +16,12 @@ import { copyFile, copyFileSync } from "ext:deno_node/_fs/_fs_copy.ts";
 import { cp, cpSync } from "ext:deno_node/_fs/_fs_cp.ts";
 import Dir from "ext:deno_node/_fs/_fs_dir.ts";
 import { exists, existsSync } from "ext:deno_node/_fs/_fs_exists.ts";
-import { fchmod, fchmodSync } from "ext:deno_node/_fs/_fs_fchmod.ts";
 import { fchown, fchownSync } from "ext:deno_node/_fs/_fs_fchown.ts";
-import { fdatasync, fdatasyncSync } from "ext:deno_node/_fs/_fs_fdatasync.ts";
 import { fstat, fstatSync } from "ext:deno_node/_fs/_fs_fstat.ts";
-import { fsync, fsyncSync } from "ext:deno_node/_fs/_fs_fsync.ts";
 import { ftruncate, ftruncateSync } from "ext:deno_node/_fs/_fs_ftruncate.ts";
 import { futimes, futimesSync } from "ext:deno_node/_fs/_fs_futimes.ts";
 import { lchmod, lchmodSync } from "ext:deno_node/_fs/_fs_lchmod.ts";
 import { lchown, lchownSync } from "ext:deno_node/_fs/_fs_lchown.ts";
-import { link, linkSync } from "ext:deno_node/_fs/_fs_link.ts";
 import { lstat, lstatSync } from "ext:deno_node/_fs/_fs_lstat.ts";
 import { lutimes, lutimesSync } from "ext:deno_node/_fs/_fs_lutimes.ts";
 import { mkdir, mkdirSync } from "ext:deno_node/_fs/_fs_mkdir.ts";
@@ -44,7 +40,6 @@ import { stat, Stats, statSync } from "ext:deno_node/_fs/_fs_stat.ts";
 import { statfs, statfsSync } from "ext:deno_node/_fs/_fs_statfs.ts";
 import { symlink, symlinkSync } from "ext:deno_node/_fs/_fs_symlink.ts";
 import { truncate, truncateSync } from "ext:deno_node/_fs/_fs_truncate.ts";
-import { unlink, unlinkSync } from "ext:deno_node/_fs/_fs_unlink.ts";
 import { utimes, utimesSync } from "ext:deno_node/_fs/_fs_utimes.ts";
 import { unwatchFile, watch, watchFile } from "ext:deno_node/_fs/_fs_watch.ts";
 // @deno-types="./_fs/_fs_write.d.ts"
@@ -75,12 +70,19 @@ import {
 import { glob, globSync } from "ext:deno_node/_fs/_fs_glob.ts";
 import {
   parseFileMode,
+  validateFunction,
+  validateInt32,
   validateInteger,
   validateObject,
   validateString,
 } from "ext:deno_node/internal/validators.mjs";
 import type { Buffer } from "node:buffer";
-import { op_fs_read_file_async } from "ext:core/ops";
+import { FsFile } from "ext:deno_fs/30_fs.js";
+import {
+  op_fs_fchmod_async,
+  op_fs_fchmod_sync,
+  op_fs_read_file_async,
+} from "ext:core/ops";
 import { core, primordials } from "ext:core/mod.js";
 
 const {
@@ -88,6 +90,7 @@ const {
   ErrorPrototype,
   ObjectPrototypeIsPrototypeOf,
   PromisePrototypeThen,
+  SymbolFor,
 } = primordials;
 
 const {
@@ -357,6 +360,111 @@ function closeSync(fd: number) {
   // TODO(@littledivy): Treat `fd` as real file descriptor. `rid` is an
   // implementation detail and may change.
   core.close(fd);
+}
+
+function fchmod(
+  fd: number,
+  mode: string | number,
+  callback: CallbackWithError,
+) {
+  validateInteger(fd, "fd", 0, 2147483647);
+  mode = parseFileMode(mode, "mode");
+  callback = makeCallback(callback);
+
+  PromisePrototypeThen(
+    op_fs_fchmod_async(fd, mode),
+    () => callback(null),
+    callback,
+  );
+}
+
+function fchmodSync(fd: number, mode: string | number) {
+  validateInteger(fd, "fd", 0, 2147483647);
+
+  op_fs_fchmod_sync(fd, parseFileMode(mode, "mode"));
+}
+
+function fdatasync(
+  fd: number,
+  callback: CallbackWithError,
+) {
+  validateInt32(fd, "fd", 0);
+  PromisePrototypeThen(
+    new FsFile(fd, SymbolFor("Deno.internal.FsFile")).syncData(),
+    () => callback(null),
+    callback,
+  );
+}
+
+function fdatasyncSync(fd: number) {
+  validateInt32(fd, "fd", 0);
+  new FsFile(fd, SymbolFor("Deno.internal.FsFile")).syncDataSync();
+}
+
+function fsync(
+  fd: number,
+  callback: CallbackWithError,
+) {
+  validateInt32(fd, "fd", 0);
+  PromisePrototypeThen(
+    new FsFile(fd, SymbolFor("Deno.internal.FsFile")).sync(),
+    () => callback(null),
+    callback,
+  );
+}
+
+function fsyncSync(fd: number) {
+  validateInt32(fd, "fd", 0);
+  new FsFile(fd, SymbolFor("Deno.internal.FsFile")).syncSync();
+}
+
+function link(
+  existingPath: string | Buffer | URL,
+  newPath: string | Buffer | URL,
+  callback: CallbackWithError,
+) {
+  existingPath = getValidatedPathToString(existingPath);
+  newPath = getValidatedPathToString(newPath);
+
+  PromisePrototypeThen(
+    Deno.link(existingPath, newPath),
+    () => callback(null),
+    callback,
+  );
+}
+
+function linkSync(
+  existingPath: string | Buffer | URL,
+  newPath: string | Buffer | URL,
+) {
+  existingPath = getValidatedPathToString(existingPath);
+  newPath = getValidatedPathToString(newPath);
+
+  Deno.linkSync(existingPath, newPath);
+}
+
+function unlink(
+  path: string | Buffer | URL,
+  callback: (err?: Error) => void,
+): void {
+  path = getValidatedPathToString(path);
+  validateFunction(callback, "callback");
+
+  PromisePrototypeThen(
+    Deno.remove(path),
+    () => callback(),
+    (err: Error) =>
+      callback(denoErrorToNodeError(err, { syscall: "unlink", path })),
+  );
+}
+
+function unlinkSync(path: string | Buffer | URL): void {
+  path = getValidatedPathToString(path);
+  try {
+    Deno.removeSync(path);
+  } catch (err) {
+    throw denoErrorToNodeError(err as Error, { syscall: "unlink", path });
+  }
 }
 
 /**
