+++ ext/node/polyfills/internal/fs/promises.ts
@@ -17,7 +17,6 @@ import { copyFilePromise } from "ext:deno_node/_fs/_fs_copy.ts";
 import { cpPromise } from "ext:deno_node/_fs/_fs_cp.ts";
 import { lchmodPromise } from "ext:deno_node/_fs/_fs_lchmod.ts";
 import { lchownPromise } from "ext:deno_node/_fs/_fs_lchown.ts";
-import { linkPromise } from "ext:deno_node/_fs/_fs_link.ts";
 import { lstatPromise } from "ext:deno_node/_fs/_fs_lstat.ts";
 import { lutimesPromise } from "ext:deno_node/_fs/_fs_lutimes.ts";
 import { mkdirPromise } from "ext:deno_node/_fs/_fs_mkdir.ts";
@@ -35,7 +34,6 @@ import { statPromise } from "ext:deno_node/_fs/_fs_stat.ts";
 import { statfsPromise } from "ext:deno_node/_fs/_fs_statfs.ts";
 import { symlinkPromise } from "ext:deno_node/_fs/_fs_symlink.ts";
 import { truncatePromise } from "ext:deno_node/_fs/_fs_truncate.ts";
-import { unlinkPromise } from "ext:deno_node/_fs/_fs_unlink.ts";
 import { utimesPromise } from "ext:deno_node/_fs/_fs_utimes.ts";
 import { watchPromise } from "ext:deno_node/_fs/_fs_watch.ts";
 import {
@@ -205,6 +203,48 @@ const chownPromise = promisify(chown) as (
   path: string | Buffer | URL,
   uid: number,
   gid: number,
+) => Promise<void>;
+
+// -- link --
+
+function link(
+  existingPath: string | Buffer | URL,
+  newPath: string | Buffer | URL,
+  callback: CallbackWithError,
+) {
+  existingPath = getValidatedPathToString(existingPath);
+  newPath = getValidatedPathToString(newPath);
+
+  PromisePrototypeThen(
+    Deno.link(existingPath, newPath),
+    () => callback(null),
+    callback,
+  );
+}
+
+const linkPromise = promisify(link) as (
+  existingPath: string | Buffer | URL,
+  newPath: string | Buffer | URL,
+) => Promise<void>;
+
+// -- unlink --
+
+function unlink(
+  path: string | Buffer | URL,
+  callback: (err?: Error) => void,
+): void {
+  path = getValidatedPathToString(path);
+
+  PromisePrototypeThen(
+    Deno.remove(path),
+    () => callback(),
+    (err: Error) =>
+      callback(denoErrorToNodeError(err, { syscall: "unlink", path })),
+  );
+}
+
+const unlinkPromise = promisify(unlink) as (
+  path: string | Buffer | URL,
 ) => Promise<void>;
 
 // -- promises object --
