+++ tools/core_import_map.json
@@ -534,11 +534,8 @@
     "ext:deno_node/_fs/cp/cp_sync.ts": "../ext/node/polyfills/_fs/cp/cp_sync.ts",
     "ext:deno_node/_fs/_fs_dir.ts": "../ext/node/polyfills/_fs/_fs_dir.ts",
     "ext:deno_node/_fs/_fs_exists.ts": "../ext/node/polyfills/_fs/_fs_exists.ts",
-    "ext:deno_node/_fs/_fs_fchmod.ts": "../ext/node/polyfills/_fs/_fs_fchmod.ts",
     "ext:deno_node/_fs/_fs_fchown.ts": "../ext/node/polyfills/_fs/_fs_fchown.ts",
-    "ext:deno_node/_fs/_fs_fdatasync.ts": "../ext/node/polyfills/_fs/_fs_fdatasync.ts",
     "ext:deno_node/_fs/_fs_fstat.ts": "../ext/node/polyfills/_fs/_fs_fstat.ts",
-    "ext:deno_node/_fs/_fs_fsync.ts": "../ext/node/polyfills/_fs/_fs_fsync.ts",
     "ext:deno_node/_fs/_fs_futimes.ts": "../ext/node/polyfills/_fs/_fs_futimes.ts",
     "ext:deno_node/_fs/_fs_ftruncate.ts": "../ext/node/polyfills/_fs/_fs_ftruncate.ts",
     "ext:deno_node/_fs/_fs_glob.ts": "../ext/node/polyfills/_fs/_fs_glob.ts",
@@ -552,7 +549,6 @@
     "ext:deno_node/_fs/_fs_stat.ts": "../ext/node/polyfills/_fs/_fs_stat.ts",
     "ext:deno_node/_fs/_fs_statfs.ts": "../ext/node/polyfills/_fs/_fs_statfs.ts",
     "ext:deno_node/_fs/_fs_symlink.ts": "../ext/node/polyfills/_fs/_fs_symlink.ts",
-    "ext:deno_node/_fs/_fs_unlink.ts": "../ext/node/polyfills/_fs/_fs_unlink.ts",
     "ext:deno_node/_fs/_fs_utimes.ts": "../ext/node/polyfills/_fs/_fs_utimes.ts",
     "ext:deno_node/_fs/_fs_watch.ts": "../ext/node/polyfills/_fs/_fs_watch.ts",
     "ext:deno_node/_fs/_fs_write.ts": "../ext/node/polyfills/_fs/_fs_write.ts",
